Count every fold round against the re-read bound
A fold that reached its bound and found the entry naming a different
extent-map slot started the round count over, so a slot whose map
slot moved under every round could re-read without end -- the bound
it was meant to be under was one a round could reset.  Every round
now counts, and a second, larger ceiling bounds the rounds in all:
past it the fold refuses with an error the caller retries, rather
than looping.

§13's bmfold point takes a count, like the other counted points: it
arms the next n rounds of a fold to park, and arming also releases a
parked one, so a test can walk a fold round by round with a writer
moving the map inside each round.  That is what drives the bound to
its fallback.

+++ lib/store.h
@@ -270,8 +270,8 @@ struct Store
 	uvlong	bmnfold;		/* slots the pass has folded */
 	uvlong	bmnreread;		/* folds the stamp sent round again */
 	uvlong	bmswapped;		/* pages the last swap installed */
-	int	bmfoldhold;		/* §13's bmfold point: park one fold */
-	int	bmfoldgo;		/* ... until the hook lets it go */
+	ulong	bmfoldhold;		/* §13's bmfold point: rounds to park */
+	int	bmfoldgo;		/* ... until the hook lets one go */
 	Rendez	bmrz;			/* on qlstate: the fold parked there */
 	Sgrain	**stagebuck;		/* §6's hash of reserved grains */
 	Sgrain	*stagefree;		/* recycled nodes */
